moving BoardContext to data folder
